% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{update_formula.brmsfit}
\alias{update_formula.brmsfit}
\title{Update model formula to exclude \code{x} predictor}
\usage{
update_formula.brmsfit(object, x)
}
\arguments{
\item{object}{An object of class 'brmsfit', fitted with \code{family = logistic("logit")} or \code{family = bernoulli("logit")}.}

\item{x}{String character indicating them name of the variable or which to compute the mid-point of the logistic curve.}
}
\description{
Internal function that removes any predictor in which \code{x} (the variable for which to compute the mid-point of the logistic curve) is involved (main effects and interactions). The resulting formula will be used to compute the linear component of the numerator in the mid-point formula. For more details see \code{\link[=use_formula]{use_formula()}}.
}
\author{
Gonzalo Garcia-Castro
}
